requirements.txt
streamlit
openai
python-dotenv
speechrecognition
pyttsx3pip 
pyaudio